pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s-ai"
version = "0.1.0"
description = "CLI that explains Kubernetes pod failures from logs using AI. Works locally with kubectl."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Dheeraj Koduru" }
]

keywords = [
  "kubernetes",
  "k8s",
  "devops",
  "sre",
  "logs",
  "ai",
  "cli"
]

dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "httpx>=0.27"
]

[project.scripts]
k8s-ai = "k8s_ai.cli:app"

[tool.setuptools]
packages = ["k8s_ai"]
